test_grid_statoil split into 4 tests.
