+++ python/tests/ecl_tests/CMakeLists.txt
@@ -19,6 +19,9 @@ set(TEST_SOURCES
     test_grid.py
     test_cell.py
     test_grid_statoil.py
+    test_grid_statoil_1.py
+    test_grid_statoil_2.py
+    test_grid_statoil_3.py
     test_grid_generator.py
     test_indexed_read.py
     test_ecl_kw_statoil.py
@@ -77,6 +80,9 @@ addPythonTest(tests.ecl_tests.test_ecl_file_statoil.EclFileStatoilTest)
 addPythonTest(tests.ecl_tests.test_grdecl_statoil.GRDECLStatoilTest)
 addPythonTest(tests.ecl_tests.test_grdecl.GRDECLTest)
 addPythonTest(tests.ecl_tests.test_grid_statoil.GridTest)
+addPythonTest(tests.ecl_tests.test_grid_statoil_1.GridTest_1)
+addPythonTest(tests.ecl_tests.test_grid_statoil_2.GridTest_2)
+addPythonTest(tests.ecl_tests.test_grid_statoil_3.GridTest_3)
 addPythonTest(tests.ecl_tests.test_ecl_kw_statoil.KWTest)
 addPythonTest(tests.ecl_tests.test_ecl_init_file.InitFileTest)
 addPythonTest(tests.ecl_tests.test_ecl_restart_file.RestartFileTest)
